% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{findbyPDB}
\alias{findbyPDB}
\title{Find PDB specific interactions}
\usage{
findbyPDB(dbname, lname)
}
\arguments{
\item{Your}{ProBiS interaction file and PDB name}
}
\description{
Use this function to obtain all PDB specific interactions
}
\examples{
findbyPDB(imported_file_name e.g. mol1.lig, '3epoA')
}
\keyword{ProBiSregion}

