% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{interactionList}
\alias{interactionList}
\title{Function to display unique interactions present according to PDBs}
\usage{
interactionList(interaction_file)
}
\arguments{
\item{Your}{ProBiS interaction file}
}
\description{
Use this function to obtain uniqe PDB to PDB interactions
}
\examples{
interactionList(imported_file_name e.g. mol1.lig)
}
\keyword{ProBiSregion}

