% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{maxInteraction}
\alias{maxInteraction}
\title{Function to display single loci within most interactions}
\usage{
maxInteraction(interaction_file)
}
\arguments{
\item{Your}{ProBiS interaction file}
}
\description{
Use this function to obtain most present loci in interaction set
}
\examples{
maxInteraction(imported_file_name e.g. mol1.lig)
}
\keyword{ProBiSmaxLoci}

