% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{mostCommonInteraction}
\alias{mostCommonInteraction}
\title{Function to display most common interactions present}
\usage{
mostCommonInteraction(interaction_file)
}
\arguments{
\item{Your}{ProBiS interaction file}
}
\description{
Use this function to find loci within most interactions
}
\examples{
mostCommonInteraction(imported_file_name e.g. mol1.lig)
}
\keyword{ProBiSmaxLoci}

