% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{plotbc}
\alias{plotbc}
\title{Function to plot interaction intensities for specific loci}
\usage{
plotbc(ifile, top)
}
\arguments{
\item{Your}{ProBiS interaction file and number of desired loci}
}
\description{
Use this function to plot frequeency of loci within interaction
}
\examples{
plotbc(imported_file_name e.g. mol1.lig, 10)
}
\keyword{ProBiSplot}

