% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{predictedArea}
\alias{predictedArea}
\title{Function to display predicted area of interaction}
\usage{
predictedArea(interaction_file)
}
\arguments{
\item{Your}{ProBiS interaction file}
}
\description{
Use this function to obtain all loci within predicted PI
}
\examples{
predictedArea(imported_file_name e.g. mol1.lig)
}
\keyword{ProBiSregion}

