% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{probisSummary}
\alias{probisSummary}
\title{Entire summary of your .lig file.}
\usage{
probisSummary(interaction_file)
}
\arguments{
\item{Your}{ProBiS interaction file}
}
\description{
Use this function obtain quick insight into your .lig file
}
\examples{
probisSummary(imported_file_name e.g. mol1.lig)
}
\keyword{ProBiSsummary}

