% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{pymolRegion}
\alias{pymolRegion}
\title{Obtain raw data to paste into PyMol for plotting and selection}
\usage{
pymolRegion(ligandFile)
}
\arguments{
\item{Your}{ProBiS interaction file}
}
\description{
Paste output to selection function in PyMol
}
\examples{
pymolRegion(imported_file_name e.g. mol1.lig)
}
\keyword{ProBiSPymoL}

