% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{ValidateProBiS}
\alias{ValidateProBiS}
\title{A validation function of ProBiS .lig files}
\usage{
ValidateProBiS(file1)
}
\arguments{
\item{Your}{.lig file - a ProBiS algorithm output.}
}
\description{
Use this function to check whether your files are of correct format.
}
\examples{
ValidateProBiS(imported_file_name e.g. mol1.lig)
}
\keyword{chackFile}

